Move retrieveTransactionType to Transaction class and override methods for specific operations
